feature: Chrome Extension + Initial Sync API Endpoint Setup (#30)
## Description

Closes https://github.com/lenghuang/leetcode-review/issues/19. Sets up a
skeleton endpoint and a Chrome Extension to enumerate the submissions
API and send it over to my server.

## Testing Details

Chrome Extension
- `console.log` doesn't show, so I typically have to do `console.error`
to get it to show in `chrome://extensions/`
- If programming in `activeTab` script, things will show up in the tab's
specific console log
- Hard to track undefined, and null errors, since sometimes it just
didn't behave.

API Endpoint
- This part was pretty straight forward
- Run `npm run dev` and attempt to click the sync button
- See status logs on terminal, and see the response in network tab

## Reflections

Was able to use V0 to get a proof of concept chrome extension first,
then build on that with my own opinion on how to structure code. I got
distracted for a bit, trying to make my JS super modular. This failed
quickly because my code was still pretty coupled, and I just put it in
different files. So I decided to keep it simple and have one big js
file. Additionally, I did lots of researching on this, trying out
bookmarklets, then an extension, and think I got a good solution going.
While this took a while because other life matters got in the way, I'm
hopeful that this will work. Got some CORS funkiness I had to work out,
but nothing that I couldn't solve by setting the right headers.

## Links Referenced
-
https://nextjs.org/docs/app/building-your-application/routing/route-handlers
- https://developer.chrome.com/docs/extensions/reference/api/tabs
- https://vercel.com/templates/next.js/edge-functions-add-header
- https://github.com/jungsu-kwon/Leetcode-Downloader-for-Submissions
- https://github.com/vsavko/Leetcode-parser
- https://github.com/TheShubham99/leetcode-to-pdf
-
https://stackoverflow.com/questions/43871637/no-access-control-allow-origin-header-is-present-on-the-requested-resource-whe

## Screenshots


![proxy](https://github.com/user-attachments/assets/259509ff-3355-4b52-a5cc-85cb24605ff3)