[prototype] Set up content scripts, messaging, and a sync page (#36)
## Description

Proof of concept PR that sets up the chrome extension in a way that we
can pass messages from Recode to Leetcode and back. This will be the
foundation for sharing data between the websites without storing
sensitive credentials.

## Reflections

This code is pretty ugly and I realized I want to redesign a lot of it
-- specifically making the loading in and UI on the extension more
asynchronous. I also want to consider multiple scenarios like if someone
wanted to open this extension from either Leetcode or my website.

Running into many hard to debug situations, so I think I'll restart this
in type script and see what happens.

Most of this is not really working code. I just want to keep this pull
request for history's sake.

## Links Referenced
-
https://www.reddit.com/r/webdev/comments/150zffo/how_to_communicate_between_chrome_extension_and_a/
- https://beastx.ro/supabase-login-with-oauth-in-chrome-extensions
-
https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/API/runtime/onMessageExternal
-
https://stackoverflow.com/questions/51091356/chrome-tabs-onupdated-doesnt-consider-a-refresh-an-update
-
https://stackoverflow.com/questions/7668502/cant-connect-to-localhost-from-chrome-extension
- https://supabase.com/docs/reference/javascript/auth-refreshsession
- https://supabase.com/docs/guides/auth/sessions <-- try calling set
session on crhome extension side
- https://supabase.com/docs/reference/javascript/auth-setsession
-
https://stackoverflow.com/questions/19035159/pass-a-message-from-chrome-extension-to-webpage
- https://stackoverflow.com/a/14245504
-
https://developer.chrome.com/docs/extensions/develop/concepts/messaging#simple
- https://stackoverflow.com/a/9916089

## Screenshots 

These are some work in progress screenshots of when I finally got the
messages to communicate correctly (sorta) between the background script
and chrome extension. There are still some weird things with duplicates,
but one step at a time :P Had to record on my phone since my intel i5
mac is burning up here and would crash if I did screen record... lol


https://github.com/user-attachments/assets/2e98406e-3fde-4164-b1c5-0ff032446fd4


https://github.com/user-attachments/assets/551372eb-0148-447d-ad6b-4d73b3566007